--- a/GameData/KSPCommunityFixes/Settings.cfg
+++ b/GameData/KSPCommunityFixes/Settings.cfg
@@ -313,6 +313,14 @@ KSP_COMMUNITY_FIXES
   // Prevent performance drops when there are in-progress comet sample or rover construction contracts
   ContractProgressEnumCache = true
 
+  // Faster and more reliable implementation of drag cube generation. Improves overall loading times (both game 
+  // load and scene/vessel/ship load times), prevent occasional lag spikes (in the editor mostly) and fix some 
+  // issues causing incorrect drag cubes to be generated (notable examples are the stock inflatable heat shield,
+  // the 1.25m and 2.5m nose cones and the Mainsail shroud). Note that by design, this patch results in a small 
+  // deviation from the stock behavior for buyoancy, aerodynamics and thermodynamics, as the generated drag cubes 
+  // will be slightly different.
+  DragCubeGeneration = true
+
   // ##########################
   // Modding
   // ##########################